feat(component): rename Toggle -\> Wrapper, update props
- Wrapper은 결국 2가지의 타입인 그룹에 대한 클릭과, 블록에 대한 클릭으로 나뉘도록 설정했다.
- 결과적으로 책임을 단일적으로 하여, 로직의 복잡성을 덜어내는 데 성공했다.
