+++ packages/ui/block-group/index.ts
@@ -1,1 +1,5 @@
-export * from './BlockGroup';
+export * from './MemberList';
+export * from './Member';
+export * from './Block';
+export * from './Wrapper';
+export * from './types';
